minor changes in logics
